pyproject.toml
[build-system]
requires = [
    "hatchling>=1.27",
]
build-backend = "hatchling.build"

[project]
name = "keepachangelog-manager-fork"
description = "Python package allowing you to manage your CHANGELOG.md files"
readme = "README.md"
version = "5.0.0"
requires-python = ">=3.9"
license = "Apache-2.0"
license-files = ["LICENSE.md"]
dependencies = [
    "pyyaml>=6.0.2,<7",
    "semantic_version>=2.10.0,<3",
    "keepachangelog>=2.0.0",
    "urllib3>=2.6.3",
    "inquirer>=3.2.2",
    "tomli>=2.0.1; python_version < '3.11'",
]
[project.urls]
Homepage = "https://github.com/matthewdeanmartin/keepachangelog-manager"
Issues = "https://github.com/matthewdeanmartin/keepachangelog-manager/issues"
Source = "https://github.com/matthewdeanmartin/keepachangelog-manager"

[project.scripts]
changelogmanager = "changelogmanager.__main__:main"
keepachangelog-manager = "changelogmanager.__main__:main"

[dependency-groups]
test = [
    "freezegun",
    "hypothesis",
    "pytest",
    "pytest-cov",
    "pytest-mock",
]
lint = [
    "black",
    "flake8",
    "flake8-pyproject",
    "flake8_formatter_junit_xml",
    "mypy",
    "pylint",
    "ruff>=0.15.12",
    "types-PyYAML",
]
security = [
    "bandit",
]
dev = [
    { include-group = "test" },
    { include-group = "lint" },
    { include-group = "security" },
]

[tool.mypy]
python_version = "3.9"
strict = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
show_error_codes = true

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 160
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "RET", "ARG", "PTH", "ERA"]
ignore = [
    "E203", "E501", # line length handled by line-length
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "F811", "ARG001", "ARG002", "ARG005", "A004"]
"changelogmanager/_llvm_diagnostics/*" = ["N818", "A001"]

[tool.hatch.build.targets.wheel]
packages = ["changelogmanager"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["changelogmanager"]

[tool.flake8]
exclude = [
    ".cache",
    ".eggs",
    ".git",
    ".tox",
    "__pycache__",
    "doc/source/conf.py",
    "venv*",
    "*.pyc",
    "*.egg-info",
]
ignore = [
    "E121",
    "E123",
    "E126",
    "E133",
    "E203",
    "E221",
    "E226",
    "E241",
    "E242",
    "E251",
    "E704",
    "W503",
    "W504",
    "W505",
]
max-line-length = 160
